app/fastapi/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
scikit-surprise
numpy<2.0
scikit-learn
pandas